rename package parser -> qparser
