+++ router/qparser/comment_test.go
@@ -1,4 +1,4 @@
-package parser
+package qparser
 
 import (
 	"fmt"
